extract duel gathering slots time in separate lib module
